feat(gptme-sessions): make --since support sub-day windows and natural phrasing
`--since` previously parsed only whole days: `2h` was rounded up to a minimum
of 1 day (math.ceil(2/24) -> 1), so `query --since 2h` silently returned a
full day of records, and inputs like `30m` or `2 hours ago` errored outright.

Rework `_parse_since` to return fractional days and accept:
- compact units: `90s`, `30m`, `2h`, `7d`, `2w` (s/m/h/d/w; `m` = minutes)
- bare numbers as days (back-compat: `30` == `30d`)
- natural phrases: `"2 hours ago"`, `"30 minutes"`, `"1 week ago"`
- `all` (no filter), case- and whitespace-insensitive

`store.query` already multiplies `since_days` by 86400, so sub-day windows now
filter on timestamps with hour/minute precision. Day-granular discovery paths
(`discover`/`sync`/unsynced hints) ceil sub-day windows to whole days, so their
file-by-date scan is unchanged. Help text and README updated.

Adds 24 unit tests (units, phrases, fractional granularity, an end-to-end
`--since 2h` filter excluding a 5h-old record, and invalid-input rejection).

+++ packages/gptme-sessions/src/gptme_sessions/store.py
@@ -187,7 +187,7 @@ class SessionStore:
         category: str | None = None,
         harness: str | None = None,
         outcome: str | None = None,
-        since_days: int | None = None,
+        since_days: float | None = None,
         project: str | None = None,
     ) -> list[SessionRecord]:
         """Filter session records by criteria."""
